lint(gui): split up state types
