+++ gui/src/state/blend.ts
@@ -1,0 +1,21 @@
+import {
+  BlendParams,
+  BrushParams,
+  ModelParams,
+  UpscaleParams,
+} from '../types/params.js';
+import { TabState } from './types.js';
+
+export interface BlendSlice {
+  blend: TabState<BlendParams>;
+  blendBrush: BrushParams;
+  blendModel: ModelParams;
+  blendUpscale: UpscaleParams;
+
+  resetBlend(): void;
+
+  setBlend(blend: Partial<BlendParams>): void;
+  setBlendBrush(brush: Partial<BrushParams>): void;
+  setBlendModel(model: Partial<ModelParams>): void;
+  setBlendUpscale(params: Partial<UpscaleParams>): void;
+}
