+++ gui/src/state/reset.ts
@@ -1,0 +1,3 @@
+export interface ResetSlice {
+  resetAll(): void;
+}
